Fix issue the deleted item is selected again with batch delete
Items recently deleted with batch action are shown in deletion
confirmation dialog when execute the batch delete action again.

Steps to reproduce:
1. Create 2 new items
2. Delete one of them using the batch delete action on top of the table
3. Select the another item and click the batch delete action

Expect to see the confirmation dialog shows one selected item,
but it shows the item that was deleted in step 2 as well.
Also, HTTP DELETE method for already deleted item is submitted
again and occurs error.

This patch fixes this issue.

Change-Id: I6555f3d2a4c0f9a4fb9f3fa780f31e3d2b03f34e
Closes-Bug: #1742599