fix tox python3 overrides
We want to default to running all tox environments under python 3, so
set the basepython value in each environment.

We do not want to specify a minor version number, because we do not
want to have to update the file every time we upgrade python.

We do not want to set the override once in testenv, because that
breaks the more specific versions used in default environments like
py35 and py36.

To install horizon, check python3 version and set target directory
in postinstall script on package.json.

Change-Id: I2592e903b51ef5bbb3672cdc036376d3117ff197
Co-Authored-By: Shu Muto <shu.mutow@gmail.com>